// superset-frontend/plugins/plugin-chart-parking-spot-matrix/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import { QueryFormData, TimeseriesDataRecord } from '@superset-ui/core';

export interface ParkingSpotMatrixStylesProps {
  height: number;
  width: number;
}

export interface ParkingSpotMatrixCustomizeProps {
  // Column Mappings
  spotIdColumn: string;
  workOrderColumn?: string;
  priorityColumn?: string;
  dueDateColumn?: string;
  stageTagColumn?: string;
  stageColorColumn?: string;
  binC1Column?: string;
  binC2Column?: string;
  binC3Column?: string;
  binBCColumn?: string;

  // Grid / Layout settings
  columnsCount: number;
  cardGap: number;
  cardBgColor: string;
  cardBorderColor: string;
  headerBgColor: string;
  headerTextColor: string;
  stageBadgeBgColor: string;
  stageBadgeTextColor: string;
  priorityBadgeBgColor: string;
  priorityBadgeTextColor: string;
  fontSize: number;
  showBinGrid: boolean;
  emptySpotText: string;
}

export interface ParkingSpotMatrixProps extends ParkingSpotMatrixStylesProps {
  data: TimeseriesDataRecord[];
  customize: ParkingSpotMatrixCustomizeProps;
}

export type ParkingSpotMatrixQueryFormData = QueryFormData &
  ParkingSpotMatrixStylesProps &
  ParkingSpotMatrixCustomizeProps;